Create script.js
first of all, all the variables are accessed using the query selected Then a conversion of the pound to kg is added with an if condition to check if the user chose pound as a weight unit. then conversion of height from feet and inches into meters After that BMI is calculated using a formula now with four if else conditions. we check the BMI value and show the status of underweight overweight or normal on the screen